# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateReviewRequest:
    properties:
      comment:
        example: Great location and very clean.
        type: string
      rating:
        example: 4
        maximum: 5
        minimum: 1
        type: integer
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  handlers.MessageResponse:
    properties:
      data:
        additionalProperties:
          type: string
        type: object
      success:
        example: true
        type: boolean
    type: object
...
  handlers.ReviewListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.ReviewResponse:
    properties:
      data:
        $ref: '#/definitions/models.Review'
      success:
        example: true
        type: boolean
    type: object
  handlers.SaveSearchHistoryRequest:
    properties:
      search_term:
        example: apartment Douala
        type: string
    type: object
  handlers.SearchHistoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.SearchHistory'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.SearchSuggestionsResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_verified:
        type: boolean
      property_id:
        type: integer
      rating:
        description: 1-5
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.SearchHistory:
    properties:
      created_at:
        type: string
      id:
        type: integer
      search_term:
        type: string
      user_id:
        type: integer
    type: object
  models.User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      locale:
        type: string
      member_since:
        type: string
      phone_number:
        type: string
      profile_picture:
        type: string
      timezone:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@gidana.com
    name: Gidana Support
  description: Real-estate platform API — property listings, rentals, messaging, digital
    wallets & payments.
  title: Gidana API
  version: 1.0.0
paths:
...
  /properties/{id}/reviews:
    get:
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReviewListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List all reviews for a property
      tags:
      - reviews
    post:
      consumes:
      - application/json
      parameters:
      - description: Property ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.ReviewResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit a review for a property
      tags:
      - reviews
...
  /reviews/{id}:
    delete:
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - reviews
  /search/history:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Clear the current user's search history
      tags:
      - search
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SearchHistoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the current user's search history
      tags:
      - search
    post:
      consumes:
      - application/json
      parameters:
      - description: Search term
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.SaveSearchHistoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
      summary: Save a search term to history (no-op if not authenticated)
      tags:
      - search
  /search/suggestions:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SearchSuggestionsResponse'
      summary: Get search auto-complete suggestions
      tags:
      - search
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"